Can get a collection of Classifiables w/ 'tests'.
Function claims it's by concept string but for now we're ignoring that.
Query needs to be expanded so the concept string (and classifier) is
returned as well for each classifiable to match the class definition. The
test is not proper for this function since it's not running on a proper
test dataset, but on whatever database is currently running.