Cleaned up some queries, added some xml comments.
Not the greatest queries, but should do for now. Need the new data in the
DummyData directory for these tests to pass. The tests should give an idea
on how to use what is done so far. Plans to make better testing setup and
be able to query the BCC for a certain term to a certain depth is next.