Connected BCC in DB to front end.
Expanded Kim's example into recursion because a term with its children
doesn't know how large it is unfortunately. Used a depth of 2 for demo
purposes. Would like to try more depth (ie most of the tree if reasonable?)
or dynamically fetch more as needed, which might be better as the BCC
expands, I would imagine.