Can addClassifiable, deleteClassifiable.
REQUIRES NEW GRAPH DB? Might be a good idea tp update!
Requires a Classifier with an email. They will get added or merged into a
Classifier with that email in the DB, so shouldn't have to worry if
whether they've been added or not. Do not call this to update, because
strange things will happen...Which is what deleteClassifiable is for I
suppose.