addClassifiable handles non-Terms in the BCC.
It complains and tells you that not all Terms in the proposed Concept
String are proper. Can be modified to add those that are, but I'm not sure
what the client would want...or what we want. But this makes it easier to
deal with for now? Ths function created to help might not be the greatest
but it should do for now.