getClassifier(email) and GLAMs don't need a url.
Thought it was a good idea to have a default URL for a GLAM.
We can add it if we need it, but the original idea was for using that
as a url if there wasn't one enetered when we add a Classifiable.
Dunno. If we do need it, holler at me and I'll fix it.