Hard coded email.
When adding/viewing your classifiables, it's using a hard coded email so
you can see how it might look. Some of the listboxes for viewing the items
looks a bit wonky with really short names, so those should probably be
cleaned up to look at least consistent with the other listboxes on the
page.