updateClassifiable - simple update works.
But then I realized that when returning a classfiable with no ConStr
would give a constr of "()". Yet, calling getClassifiableById would return
the expected constr of "". So, will be changing add to call
getClassifiableById and remove the query that would gather terms and such.
In addition, getClassifiableById needs to be updated to return ALL
relevant information. (Owner, whatever else is missing).