Tables for Classifiables are now on GLAMClass.aspx
They also won't stretch the page if the content it too long. Unfortunately,
the page had to be reorganized since the tables have so many columns.
Maybe we could have kept with the division, but it might have looked weird.
If someone wants to fix it, please go ahead! Also, not sure where the Add,
Edit and Remove buttons should go now...